requirements.txt
Django==4.2.13
Pillow==9.5.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
